Turning on warnings as error in Visual Studio
Disabled a few warnings and fixed a few warnings to enable
turning 'warnings as errors' on in Visual Studio.

Reviewed-on: https://swiftshader-review.googlesource.com/5710
Tested-by: Alexis Hétu <sugoi@google.com>

+++ src/OpenGL/libEGL/libEGL.cpp
@@ -989,7 +989,7 @@ EGLImageKHR CreateImageKHR(EGLDisplay dpy, EGLContext ctx, EGLenum target, EGLCl
 		}
 	#endif
 
-	GLuint name = reinterpret_cast<intptr_t>(buffer);
+	GLuint name = reinterpret_cast<GLuint>(buffer);
 
 	if(name == 0)
 	{
